feat: pretty printing without color (#16)
Introduce a new flag `--color` that decides whether colors are printed
while pretty-printing is on. Valid values are `on`, `off` and `auto`,
which is the default. In the future, there is the possibility of
autodetecting from TTY settings, but at the time of writing `auto`
means `on`.

The flags `-c` and `-C` can also be used to toggle color on/off respectively.

NB: Currently, only applies to EDN printing.

Additionally tests the pretty-printing with multiple line case, and
ANSI color escape sequences.